additional fixes
change tracking area after layout
